requirements.txt
flake8
pandas
python_dotenv
dropbox
streamlit
streamlit-calendar
streamlit-navigation-bar
